YH-1376: change files structure for FSD
