+++ src/entities/mostDifficultQuestions/index.ts
@@ -1,1 +1,2 @@
 export { MostDifficultQuestions } from './ui/MostDifficultQuestions';
+export { difficultQuestionsHandler } from './api/__mocks__/index';
