+++ src/entities/mostDifficultQuestions/ui/MostDifficultQuestions.module.css
@@ -1,8 +1,3 @@
-.card {
-  gap: 20px;
-  max-width: 455px;
-}
-
 .question {
   position: relative;
   margin-left: 15px;
@@ -41,11 +36,6 @@
 }
 
 @media (width <= 767px) {
-  .card {
-    padding: 16px;
-    max-width: 344px;
-  }
-
   .question span {
     white-space: wrap;
     max-width: 225px;
@@ -62,12 +52,5 @@
   .percent {
     display: flex;
     align-items: center; 
-  }
-}
-
-@media (width <= 480px) {
-  .card {
-    padding: 20px 10px;
-    max-width: 340px;
   }
 }
