+++ src/widgets/question/DifficultQuestionsList/index.ts
@@ -1,0 +1,1 @@
+export { DifficultQuestionsList } from './ui/DifficultQuestionsList';
